backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
audioread==3.0.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
dtwalign==0.1.1
eng_to_ipa==0.0.2
fastapi==0.116.1
filelock==3.19.1
fonttools==4.59.1
fsspec==2025.7.0
h11==0.16.0
hf-xet==1.1.8
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.9
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
safetensors==0.6.2
scipy==1.16.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
starlette==0.47.2
sympy==1.14.0
tokenizers==0.21.4
torch==2.8.0
torchaudio==2.8.0
tqdm==4.67.1
transformers==4.55.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
